requirements.txt
colorama==0.4.6
Pillow==10.4.0
psutil==5.9.8
python-dotenv==1.0.1
pytz==2024.2
requests==2.31.0
schedule==1.2.1
word2number==1.1

python-telegram-bot>=20.0,<21.0

pyTelegramBotAPI==4.15.4
wcwidth>=0.2.6

discord.py==2.3.2
discord-webhook==1.1.0
PyNaCl>=1.5.0,<1.6.0

aiohttp>=3.9.0
async_lru>=2.0.4
oauthlib>=3.2.2
tweepy==4.16.0

deepl==1.15.0
deep-translator==1.11.4
google-cloud-translate>=3.0.2
langdetect==1.0.9
